Refactor Fix storybook webpack config
