+++ .storybook/webpack.config.js
@@ -1,7 +1,11 @@
-const fs = require('fs')
-const ts = require('typescript')
-const tsConfig = require('./../tsconfig.json')
+const svelteBuild = require('./../configs/svelte-build')
+
 module.exports = ({ config, mode }) => {
+  // remove original storybook webpack rules for svelte
+  config.module.rules = config.module.rules.filter(r => {
+    return !r.test.toString().includes('svelte')
+  })
+
   config.module.rules.push({
     test: /\.(ts|tsx)$/,
     use: [
@@ -10,6 +14,22 @@ module.exports = ({ config, mode }) => {
       }
     ],
   });
+
+  config.module.rules.push({
+    test: /\.svelte$/,
+    exclude: /node_modules/,
+    use: {
+      loader: 'svelte-loader',
+      options: {
+        skipIntroByDefault: true,
+        nestedTransitions: true,
+        emitCss: true,
+        hotReload: true,
+        preprocess: svelteBuild.preprocess
+      }
+    }
+  })
+
   config.resolve.extensions.push('.ts', '.tsx');
   return config;
 };
